+ Moved all combat status enums to CombatStatus.h and added CombatStatus.cpp. + Renamed STATUS_TYPE to COMBAT_STATUS. + Added functions to return text from combat status and effect bits. + Added EQUIP_TECHNOLOGY enum which is going to be used for determining wether EMP affects a weapon or not. + Added COMBAT_STATUS_BLACKOUT which is now caused by EMP weapons. + Relaxed difficulty a bit because it took forever to reach higher levels. * Added ATTACK_EFFECT_LEECH property to Vampire profession. - Changed most uint32_t to int32_t in order to avoid some underflow problems.

+++ kitsulib/Inventory.h
@@ -10,9 +10,9 @@ namespace klib
 {
 	struct SItem
 	{
-		int32_t Index;
-		int32_t Modifier;
-		int32_t Level;
+		int16_t Index;
+		int16_t Modifier;
+		int16_t Level;
 	};
 
 	struct SInventorySlot
